plotting-service/plotting_service/utils.py
from pathlib import Path


def find_file(ceph_dir: str, instrument: str, experiment_number: int, filename: str) -> Path | None:
    # Run normal check
    basic_path = Path(ceph_dir) / f"{instrument.upper()}/RBNumber/RB{experiment_number}/autoreduced/{filename}"
    if basic_path.exists():
        return basic_path

    # Attempt to find file in autoreduced folder
    autoreduced_folder = Path(ceph_dir) / f"{instrument.upper()}/RBNumber/RB{experiment_number}/autoreduced"
    if autoreduced_folder.exists():
        found_paths = list(autoreduced_folder.rglob(filename))
        if len(found_paths) > 0 and found_paths[0].exists():
            return found_paths[0]

    return None
